made api modular
- Made `api` *modular*
- Finished `update | edit` logic
- Adding `add` logic
