+++ .prettierignore
@@ -1,0 +1,10 @@
+# Auto-generated by the Prettier Extension for Replit
+# node_modules is ignored by default
+.config
+.build
+.cache
+.upm
+.tmp
+.git
+replit.nix
+.replit
