+++ pyproject.toml
@@ -11,6 +11,9 @@ pyairtable = "^2.3.3"
 bcrypt = "^4.1.3"
 flask-login = "^0.6.3"
 logger = "^1.4"
+dropbox = "^12.0.2"
+werkzeug = "^3.0.3"
+api = "^0.0.7"
 
 [tool.pyright]
 # https://github.com/microsoft/pyright/blob/main/docs/configuration.md
