+++ replit.nix
@@ -1,5 +1,6 @@
 {pkgs}: {
   deps = [
+    pkgs.nodePackages.prettier
     pkgs.libxcrypt
     pkgs.rustc
     pkgs.libiconv
