+++ templates/all_companies.html
@@ -3,77 +3,92 @@
 
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>All Companies (for testing)</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        var tailwind = {
-            theme: {
-                colors: {
-                    content: '#386641',
-                    'logo-prime': '#6A994E',
-                    'logo-background': '#A7C957',
-                    background: '#E9F5DB',
-                    black: '#001000',
-                },
-            },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('tailwindcss-container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+      var tailwind = {
+        theme: {
+          colors: {
+            content: "#386641",
+            "logo-prime": "#6A994E",
+            "logo-background": "#A7C957",
+            background: "#E9F5DB",
+            black: "#001000",
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("tailwindcss-container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
+  </head>
 
-<body class="flex flex-col items-center justify-start min-h-screen">
+  <body class="flex flex-col items-center justify-start min-h-screen">
     <!-- Header -->
-    {% set nav_links = [
-        { "href": "/login", "name": "Login" },
-        { "href": "/sign-up", "name": "Sign up" },
-    ] %}
-
-    {% set logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %}
-
-    {{ navbar(logo=logo | safe, navlinks=nav_links) }}
+    {% set nav_links = [ { "href": "/login", "name": "Login" }, { "href":
+    "/sign-up", "name": "Sign up" }, ] %} {% set logo = '<span
+      class="text-3xl font-bold text-gray-800"
+      >Veric</span
+    >' %} {{ navbar(logo=logo | safe, navlinks=nav_links) }}
 
     <!-- Main Content -->
-    <main class="w-full flex flex-col items-center justify-center p-8 space-y-8 lg:px-12">
-        <div class="space-y-8 w-full">
-            {% for company in company_records %}
-                <div class="shadow-sm flex flex-col items-start justify-center p-4 space-y-2 border border-gray-300 rounded-lg bg-white w-4/5">
-                    <code class="text-sm flex flex-col font-monospace p-2 rounded max-w-full">
-                        {
-                            {% for key, value in company.items() %}
-                                <div class='flex flex-row items-start justify-start space-x-2 ml-6 py-1 border-gray-300 {% if not loop.last %}border-b{% endif %} select-auto'>
-                                    <span>"{{ key }}"</span>: <span>{{ value | tojson | safe }}{% if not loop.last %},{% endif %}</span>
-                                </div>
-                            {% endfor %}
-                        }
-                    </code>
-                </div>
-            {% endfor %}
+    <main
+      class="w-full flex flex-col items-center justify-center p-8 space-y-8 lg:px-12"
+    >
+      <div class="space-y-8 w-full">
+        {% for company in company_records %}
+        <div
+          class="shadow-sm flex flex-col items-start justify-center p-4 space-y-2 border border-gray-300 rounded-lg bg-white w-4/5"
+        >
+          <code
+            class="text-sm flex flex-col font-monospace p-2 rounded max-w-full"
+          >
+            { {% for key, value in company.items() %}
+            <div
+              class="flex flex-row items-start justify-start space-x-2 ml-6 py-1 border-gray-300 {% if not loop.last %}border-b{% endif %} select-auto"
+            >
+              <span>"{{ key }}"</span>:
+              <span
+                >{{ value | tojson | safe }}{% if not loop.last %},{% endif
+                %}</span
+              >
+            </div>
+            {% endfor %} }
+          </code>
         </div>
+        {% endfor %}
+      </div>
     </main>
 
     <!-- Footer -->
-    <footer class="h-40 flex flex-col items-center justify-center w-full space-y-4">
-        <img src='../static/images/logo-black.svg' alt="Logo" class='w-16'/>
-        <p class="text-center text-gray-700">
-            &copy; 2023 Veric. All rights reserved.
-        </p>
+    <footer
+      class="h-40 flex flex-col items-center justify-center w-full space-y-4"
+    >
+      <img src="../static/images/logo-black.svg" alt="Logo" class="w-16" />
+      <p class="text-center text-gray-700">
+        &copy; 2023 Veric. All rights reserved.
+      </p>
     </footer>
-</body>
+  </body>
 </html>
