+++ templates/applications.html
@@ -3,77 +3,90 @@
 
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Veric</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        tailwind.config = {
-            theme: {
-                extend: {
-                    colors: {
-                        content: '#386641',
-                        'logo-prime': '#6A994E',
-                        'logo-background': '#A7C957',
-                        background: '#E9F5DB',
-                        black: '#001000',
-                    },
-                },
+      tailwind.config = {
+        theme: {
+          extend: {
+            colors: {
+              content: "#386641",
+              "logo-prime": "#6A994E",
+              "logo-background": "#A7C957",
+              background: "#E9F5DB",
+              black: "#001000",
             },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('@tailwindcss/container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("@tailwindcss/container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
+  </head>
 
-<body class="flex flex-col items-center justify-start min-h-screen bg-background text-content">
-
+  <body
+    class="flex flex-col items-center justify-start min-h-screen bg-background text-content"
+  >
     <!-- Header -->
-    {% set nav_links = [
-        { "href": "/login", "name": "Login" },
-        { "href": "/sign-up", "name": "Sign up" },
-    ] %}
-
-    {% set logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %}
-
-    {{ navbar(logo=logo|safe, navlinks=nav_links) }}
+    {% set nav_links = [ { "href": "/login", "name": "Login" }, { "href":
+    "/sign-up", "name": "Sign up" }, ] %} {% set logo = '<span
+      class="text-3xl font-bold text-gray-800"
+      >Veric</span
+    >' %} {{ navbar(logo=logo|safe, navlinks=nav_links) }}
 
     <!-- Main Content -->
     <main class="flex-grow w-full px-4 py-8">
-        {% for application in applications %}
-            <div class="bg-white shadow-md rounded-lg p-4 mb-4">
-                {% for key, value in application.items() %}
-                    <div class="flex flex-row items-start justify-start space-x-2 py-1 border-gray-300 {% if not loop.last %}border-b{% endif %} select-auto">
-                        <span class="font-semibold">{{ key }}</span>: <span>{{ value | tojson | safe }}</span>
-                    </div>
-                {% endfor %}
-            </div>
+      {% for application in applications %}
+      <div class="bg-white shadow-md rounded-lg p-4 mb-4">
+        {% for key, value in application.items() %}
+        <div
+          class="flex flex-row items-start justify-start space-x-2 py-1 border-gray-300 {% if not loop.last %}border-b{% endif %} select-auto"
+        >
+          <span class="font-semibold">{{ key }}</span>:
+          <span>{{ value | tojson | safe }}</span>
+        </div>
         {% endfor %}
+      </div>
+      {% endfor %}
     </main>
 
     <!-- Footer -->
-    <footer class="h-40 flex flex-col items-center justify-center w-full space-y-4 bg-logo-background text-black">
-        <img src='../static/images/logo-black.svg' alt="Logo" class='w-16'/>
-        <p class="text-center text-gray-700">
-            &copy; 2023 Veric. All rights reserved.
-        </p>        
+    <footer
+      class="h-40 flex flex-col items-center justify-center w-full space-y-4 bg-logo-background text-black"
+    >
+      <img src="../static/images/logo-black.svg" alt="Logo" class="w-16" />
+      <p class="text-center text-gray-700">
+        &copy; 2023 Veric. All rights reserved.
+      </p>
     </footer>
 
     <!-- External JavaScript -->
-    <script src="https://kit.fontawesome.com/68c16b630a.js" crossorigin="anonymous"></script> <!-- Include Font Awesome -->
-</body>
+    <script
+      src="https://kit.fontawesome.com/68c16b630a.js"
+      crossorigin="anonymous"
+    ></script>
+    <!-- Include Font Awesome -->
+  </body>
 </html>
