+++ templates/form.html
@@ -1,86 +1,113 @@
 {% from 'macros.html' import render_form, navbar %}
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Veric</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        var tailwind = {
-            theme: {
-                colors: {
-                    content: '#386641',
-                    'logo-prime': '#6A994E',
-                    'logo-background': '#A7C957',
-                    background: '#E9F5DB',
-                    black: '#001000',
-                },
+      tailwind.config = {
+        theme: {
+          extend: {
+            colors: {
+              content: "#386641",
+              "logo-prime": "#6A994E",
+              "logo-background": "#A7C957",
+              background: "#E9F5DB",
+              black: "#001000",
             },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('tailwindcss-container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("tailwindcss-container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
+  </head>
 
-<body class="flex flex-col items-center justify-start min-h-screen">
-
+  <body class="flex flex-col items-center justify-start min-h-screen">
     <!-- Header -->
-    {% set nav_links = [
-        { "href": "/", "name": "home" },
-        { "href": "/logout", "name": "Logout" }
-    ] %}
-
-    {% set logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %}
-
-    {{ navbar(logo=logo|safe, navlinks=nav_links) }}
+    {% set nav_links = [ { "href": "/", "name": "home" }, { "href": "/logout",
+    "name": "Logout" } ] %} {% set logo = '<span
+      class="text-3xl font-bold text-gray-800"
+      >Veric</span
+    >' %} {{ navbar(logo=logo|safe, navlinks=nav_links) }}
 
     <main>
-      <form class='flex flex-col items-center justify-center p-4 space-y-2 border border-gray-300 rounded' method="post" action='/update/{{table}}/{{data.id}}'>
+      <form
+        class="flex flex-col items-center justify-center p-4 space-y-2 border border-gray-300 rounded"
+        method="post"
+        action="/update/{{ table }}/{{ data.id }}"
+        enctype="multipart/form-data"
+      >
         {% for field in formdata %}
-          {% if field.type == 'image' %}
-          <label for="{{data.fields[field.key][0]['id']}}">Select a file
-          <input type="file" id="{{data.fields[field.key][0]['id']}}" name="{{field.key}}"/>
-            </label>
+        <label for="{{ field.key }}">
+          {{ field.name }} {% if field.type == 'image' %}
+          <input
+            type="file"
+            id="{{ field.key }}"
+            name="{{ field.key }}"
+            class="mt-1 block w-full"
+          />
           {% else %}
-          <label for="{{field.key}}">
-            {{field.name}}
-            <input
-              name="{{field.key}}"
-              id="{{field.key}}"
-              type="{{field.type}}"
-              {% if field.editable %}
-                placeholder="{{data.fields[field.key]}}" 
-              {% else %} 
-                readonly = "true"
-                value="{{data.fields[field.key]}}" 
-              {% endif %}
-            />
-          </label>
+          <input
+            name="{{ field.key }}"
+            id="{{ field.key }}"
+            type="{{ field.type }}"
+            class="mt-1 block w-full"
+            {%
+            if
+            field.editable
+            %}
+            placeholder="{{ data.fields[field.key] }}"
+            {%
+            else
+            %}
+            readonly="true"
+            value="{{ data.fields[field.key] }}"
+            {%
+            endif
+            %}
+          />
           {% endif %}
+        </label>
         {% endfor %}
-        <button type="submit">Submit</button>
+        <button
+          type="submit"
+          class="mt-4 bg-blue-500 text-white py-2 px-4 rounded-lg"
+        >
+          Submit
+        </button>
       </form>
     </main>
+
     <!-- Footer -->
-    <footer class="h-40 flex flex-col items-center justify-center w-full space-y-4">
-        <img src='../static/images/logo-white.svg' alt="Logo" class='w-16'/>
-        <p class="text-center text-gray-700">
-            &copy; {{current_year}} Veric. All rights reserved.
-        </p>
+    <footer
+      class="h-40 flex flex-col items-center justify-center w-full space-y-4"
+    >
+      <img src="../static/images/logo-white.svg" alt="Logo" class="w-16" />
+      <p class="text-center text-gray-700">
+        &copy; {{ current_year }} Veric. All rights reserved.
+      </p>
     </footer>
-</body>
+  </body>
 </html>
