+++ templates/login.html
@@ -1,63 +1,65 @@
 <!-- Import Jinja Macros -->
 {% from 'macros.html' import navbar, call_to_actions as actions, render_form %}
 
-
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Veric</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        var tailwind = {
-            theme: {
-                colors: {
-                    content: '#386641',
-                    'logo-prime': '#6A994E',
-                    'logo-background': '#A7C957',
-                    background: '#E9F5DB',
-                    black: '#001000',
-                },
-            },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('tailwindcss-container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+      var tailwind = {
+        theme: {
+          colors: {
+            content: "#386641",
+            "logo-prime": "#6A994E",
+            "logo-background": "#A7C957",
+            background: "#E9F5DB",
+            black: "#001000",
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("tailwindcss-container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
+  </head>
 
-<body class="flex flex-col items-center justify-start min-h-screen">
-
+  <body class="flex flex-col items-center justify-start min-h-screen h-screen">
     <!-- Header -->
-    {% set nav_links = [
-        { "href": "/sign-up", "name": "Sign up" },
-    ] %}
-
-    {% set logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %}
-
-    {{ navbar(logo=logo|safe, navlinks=nav_links) }}
+    {% set nav_links = [ { "href": "/sign-up", "name": "Sign up" }, ] %} {% set
+    logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %} {{
+    navbar(logo=logo|safe, navlinks=nav_links) }}
     <!-- Main Content -->
-    <main class="h-full w-full absolute mt-[8rem]">
-        <div class="my-auto mx-auto p-[8rem] border border-gray rounded-2xl absolute shadow-md">
-            <h1 class="text-3xl mb-[3rem] mx-auto">Log In</h1>
-            {{ render_form(url_for('login'), 'POST', [
-                {'id': 'email', 'name': 'email', 'type': 'email', 'placeholder': 'Email', 'required': True},
-                {'id': 'password', 'name': 'password', 'type': 'password', 'placeholder': 'Password', 'required': True},
-            ]) }}
-        </div>
+    <main class="h-screen w-screen relative mt-[4rem]">
+      <div
+        class="w-[24rem] h-[24rem] mx-auto p-[6rem] border border-gray rounded-2xl shadow-md box-border"
+      >
+        <h1 class="text-3xl mb-[2rem] text-gray-800 mx-auto">Log In</h1>
+        {{ render_form(url_for('login'), 'POST', [ {'id': 'email', 'name':
+        'email', 'type': 'email', 'placeholder': 'Email', 'required': True},
+        {'id': 'password', 'name': 'password', 'type': 'password',
+        'placeholder': 'Password', 'required': True}, ]) }}
+      </div>
     </main>
-</body>
+  </body>
 </html>
