+++ templates/look_hires.html
@@ -1,52 +1,56 @@
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Veric</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        var tailwind = {
-            theme: {
-                colors: {
-                    content: '#386641',
-                    'logo-prime': '#6A994E',
-                    'logo-background': '#A7C957',
-                    background: '#E9F5DB',
-                    black: '#001000',
-                },
-            },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('tailwindcss-container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+      var tailwind = {
+        theme: {
+          colors: {
+            content: "#386641",
+            "logo-prime": "#6A994E",
+            "logo-background": "#A7C957",
+            background: "#E9F5DB",
+            black: "#001000",
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("tailwindcss-container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
-<body>
-<!--- HEADER --->
+  </head>
+  <body>
+    <!--- HEADER --->
 
+    <!--- MAIN --->
 
+    <!--- FOOTER --->
 
-<!--- MAIN --->
-
-
-
-
-<!--- FOOTER --->
-
-
-<script src="https://kit.fontawesome.com/68c16b630a.js" crossorigin="anonymous"></script> <!-- Include Font Awesome -->
-</body>
+    <script
+      src="https://kit.fontawesome.com/68c16b630a.js"
+      crossorigin="anonymous"
+    ></script>
+    <!-- Include Font Awesome -->
+  </body>
 </html>
