+++ templates/sign_up.html
@@ -1,61 +1,62 @@
 <!-- Import Jinja Macros -->
 {% from 'macros.html' import navbar, call_to_actions as actions, render_form %}
 
-
 <!DOCTYPE html>
 <html lang="en" data-theme="wireframe">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Veric</title>
 
     <!-- External CSS -->
-    <link rel="stylesheet" href="../static/css/style.css"> <!-- Include custom CSS -->
-    <script src="https://cdn.tailwindcss.com"></script> <!-- Include Tailwind CSS -->
+    <link rel="stylesheet" href="../static/css/style.css" />
+    <!-- Include custom CSS -->
+    <script src="https://cdn.tailwindcss.com"></script>
+    <!-- Include Tailwind CSS -->
 
     <!-- Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet"> <!-- Include Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+      rel="stylesheet"
+    />
+    <!-- Include Google Fonts -->
 
     <!-- Tailwind Customization Script -->
     <script>
-        var tailwind = {
-            theme: {
-                colors: {
-                    content: '#386641',
-                    'logo-prime': '#6A994E',
-                    'logo-background': '#A7C957',
-                    background: '#E9F5DB',
-                    black: '#001000',
-                },
-            },
-            plugins: [
-                require('@tailwindcss/forms'),
-                require('tailwindcss-container-queries'),
-                require('@tailwindcss/aspect-ratio')
-            ],
-        };
+      var tailwind = {
+        theme: {
+          colors: {
+            content: "#386641",
+            "logo-prime": "#6A994E",
+            "logo-background": "#A7C957",
+            background: "#E9F5DB",
+            black: "#001000",
+          },
+        },
+        plugins: [
+          require("@tailwindcss/forms"),
+          require("tailwindcss-container-queries"),
+          require("@tailwindcss/aspect-ratio"),
+        ],
+      };
     </script>
-</head>
+  </head>
 
-<body class="flex flex-col items-center justify-start min-h-screen">
-
+  <body class="flex flex-col items-center justify-start min-h-screen">
     <!-- Header -->
-    {% set nav_links = [
-        { "href": "/login", "name": "Login" },
-    ] %}
-
-    {% set logo = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %}
-
-    {{ navbar(logo=logo|safe, navlinks=nav_links) }}
+    {% set nav_links = [ { "href": "/login", "name": "Login" }, ] %} {% set logo
+    = '<span class="text-3xl font-bold text-gray-800">Veric</span>' %} {{
+    navbar(logo=logo|safe, navlinks=nav_links) }}
     <!-- Main Content -->
     <main>
-          {{ render_form(url_for('sign_up'), 'POST', [
-            {'id': 'email', 'name': 'email', 'type': 'email', 'placeholder': 'Email', 'required': True},
-            {'id': 'password', 'name': 'password', 'type': 'password', 'placeholder': 'Password', 'required': True},
-            {'id': 'company_name', 'name': 'company_name', 'type': 'text', 'placeholder': 'Company Name', 'required': True}
-          ]) }}
+      {{ render_form(url_for('sign_up'), 'POST', [ {'id': 'email', 'name':
+      'email', 'type': 'email', 'placeholder': 'Email', 'required': True},
+      {'id': 'password', 'name': 'password', 'type': 'password', 'placeholder':
+      'Password', 'required': True}, {'id': 'company_name', 'name':
+      'company_name', 'type': 'text', 'placeholder': 'Company Name', 'required':
+      True} ]) }}
     </main>
-</body>
+  </body>
 </html>
